Back to using celery for mirroring comments
One single process to send comments is not fast enough. alien.top is
already creating a huge backlog, which means that comments are not
showing up, because by the time they are processed, they are
considered to old to be worthy of replication.

To fix that and to avoid double-submission, the approach will be to
keep the single process that checks what comments should be mirrored,
and introduce a new status (accepted) to indicate that the comment
should be mirrored. Once a comment is marked as accepted, it will no
longer be checked again and it will be scheduled to be sent via a
celery task, which means that the backlog can be quickly
cleared.

To avoid flooding and too many comments being sent at the same time to
the lemmy instance, there is a small delay (currently 0.05 seconds) in
the loop that finds all comments from a given subreddit and a larger
one (1 second) after all subreddits have been checked. This means that
we will be scheduling at most 20 new mirror tasks per second, which
should be enough to avoid backlogs and also avoid issues with hitting
rate limits